Add Trainee option for Judges
Trainee option for judges added, which can be modified by administrators
only and is not a status that can be achieved through survey answers.
Functionality implemented in assignment logic that assigns trainees
preferentially if there are at least eight participants and chooses a
wing judge before the assignment of the main speakers (in order to
maximize the probability of selecting a chair as first wing judge).

Additionally, a small bug fix in the free speech preference functionality.