Improvement of debate style, fixes in BP assignment logic
Debate style is now initialized as dynamic and can be overwritten
depending on the selected scenario. If only OPD rooms are selected, it
will be finalized as OPD. This simplifies the debate creation process.

Bug fixes in BP assignment, no training mode for BP but wing selection
logic follows OPD. Wings and newbies are now treated equally for more
equal chances and more diverse juries.